Rename js modules
