+++ js/blocks/get-header-node.js
@@ -1,4 +1,4 @@
-import {INITIAL_GAME} from '../utils/changeLevel.js';
+import {INITIAL_GAME} from '../utils/change-level.js';
 
 const heartFullIcon = `<img src="img/heart__full.svg" class="game__heart" alt="Life" width="31" height="27">`;
 const heartEmptyIcon = `<img src="img/heart__empty.svg" class="game__heart" alt="Life" width="31" height="27">`;
