+++ js/main.js
@@ -1,4 +1,4 @@
-import appendNodeToMain from './utils/appendNode.js';
+import appendNodeToMain from './utils/append-node.js';
 import introNode from './blocks/intro.js';
 
 // Отрисовывает стартовый экран
