+++ js/utils/reduce-player-lives.js
